add tests

+++ src/main/java/com/tinkoff/edu/app/common/LoanRequest.java
@@ -5,7 +5,7 @@ package com.tinkoff.edu.app.common;
  */
 public class LoanRequest {
     private final int months;
-    private final int amount;
+    private final double amount;
     private final LoanType type;
     private final String fio;
 
@@ -13,7 +13,7 @@ public class LoanRequest {
         return fio;
     }
 
-    public LoanRequest(int months, int amount, LoanType type, String fio) {
+    public LoanRequest(int months, double amount, LoanType type, String fio) {
         this.months = months;
         this.amount = amount;
         this.type = type;
@@ -28,7 +28,7 @@ public class LoanRequest {
         return months;
     }
 
-    public int getAmount() {
+    public double getAmount() {
         return amount;
     }
